Extract crawl logic to crawler
