+++ lib/json_parser.rb
@@ -7,6 +7,7 @@ module JsonParser
   require 'concerns/type_cast'
   require 'concerns/options_parser'
   require 'json_parser/version'
+  require 'json_parser/crawler'
   require 'json_parser/post_processor'
   require 'json_parser/fetcher'
   require 'json_parser/class_methods'
